Corrected some of applicable code smells in soultion.

+++ PizzaShop/PizzaShop.Library/Models/Pizza.cs
@@ -65,7 +65,7 @@ namespace PizzaShop.Library
 
         public decimal CalculatePrice(decimal baseP, decimal topP)
         {
-            Price = (decimal)(baseP + topP * Toppings.Count);
+            Price = (baseP + topP * Toppings.Count);
             return Price;
         }
